requirements.txt
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.1
pytorch-lightning==1.0.8
pyyaml==5.3.1
scipy==1.5.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.50.2